stockflow-crm/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.17
python-dotenv==1.0.1
email-validator==2.2.0

# AI invoice processing — Gemini 2.5 Flash (native vision, no OCR needed)
google-genai>=1.5.0

# Email notifications
sendgrid>=6.11.0
reportlab>=4.0.0
